pyproject.toml
[tool.poetry]
name = "ptb-state-handler"
version = "0.2.0"
description = "State handler for python-telegram-bot library"
authors = ["Vlad Pastushenko <iam@vladpi.me>"]
license = "Apache License 2.0"

packages = [
    { include = "state_handler"},
]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.0"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
